Shovelling coal into the server...

+++ client/src/components/Main.js
@@ -14,6 +14,10 @@ async function createServerApi() {
 }
 
 async function deleteServerApi() {
+  const agree = window.confirm('Are you sure you want to delete your Keva instance?')
+  if (!agree) {
+    return
+  }
   const { data } = await axios.delete(`https://keva-cloud.tuhuynh.com/delete?who=${localStorage.getItem('email')}`)
   return data
 }
@@ -62,7 +66,7 @@ function Main() {
 
   return (
     <React.Fragment>
-      <h1>Hello {localStorage.getItem('email')}</h1>
+      <h1>Keva Cloud Console</h1>
       {loading && <div>Fetching your instance data...</div>}
       {!loading && !created && !health && <React.Fragment>
         <p>You haven't created any Keva instance</p>
@@ -73,9 +77,9 @@ function Main() {
         <div>Container ID: {health.container}</div>
         <div>CPU Usage: {health.cpu}</div>
         <div>Memory Usage: {health.memory.raw} ({health.memory.percent})</div>
-        <div><button onClick={() => loadHealth(true)} style={{ marginTop: '20px' }}>Refresh health</button></div>
+        <div><button className="secondary" onClick={() => loadHealth(true)} style={{ marginTop: '20px' }}>Refresh health</button></div>
         <div><a href={`https://keva-cloud.tuhuynh.com/log?id=${health.container}`} target="_blank" rel="noreferrer">
-          <button>View logs</button>
+          <button className="secondary">View logs</button>
           </a></div>
         <div><button disabled={loading} onClick={deleteServer}>Destroy this instance</button></div>
       </React.Fragment>}
@@ -88,6 +92,7 @@ function Main() {
         <div>Connect by redis-cli:</div>
         <div><code>redis-cli -h run.keva.dev -p {created.port} -a {created.pwd}</code></div>
       </React.Fragment>}
+      <div style={{ marginTop: '20px' }}>Logged in as {localStorage.getItem('email')} (logout?)</div>
     </React.Fragment>
   )
 }
